+++ tests/files/module2.py
@@ -1,10 +1,11 @@
 """
-Interative test module for modlib
+While-loop test module for modlib
 
 :author:  Walker M. White (wmw2)
 :version: June 9, 2019
 """
 
-for y in range(4):
-    x = input('Item '+str(y)+': ')
-    print('Result is '+x)
+while True:
+    pass
+
+x = 5