+++ tests/files/module3.py
@@ -1,0 +1,10 @@
+"""
+Interative test module for modlib
+
+:author:  Walker M. White (wmw2)
+:version: June 9, 2019
+"""
+
+for y in range(4):
+    x = input('Item '+str(y)+': ')
+    print('Result is '+x)