+++ tests/files/module4.py
@@ -1,0 +1,48 @@
+"""
+Problematic test module for modlib
+
+:author:  Walker M. White (wmw2)
+:version: August 9, 2019
+"""
+
+def foo():
+    while True:
+        print('x')
+    print('done')
+
+
+def duh():
+    while varb:
+        pass
+
+
+def hey():
+    print(varb)
+
+
+def bar(x):
+    return 1/x
+
+
+print('Bye')
+
+
+if __name__ == '__main__':
+    hey()
+    print('good')
+
+
+x = 1
+x = 2
+x = 3
+x = 4
+x = 5
+x = 6
+x = 7
+x = 8
+x = 9
+x = 10
+x = 11
+x = 12
+x = 13
+x = 14