+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## Version 1.2: Fall 2019 releasee
+
+* Added `assert_error` to `testcase` for checking function asserts
+* Added code rewriting to `modlib` (still unexposed) to guard against infinite loops
+* Modified `Environment` in `modlib` to load from file or source string
+* Added feature to `Environment` to execute code as as a script (`__name__ == '__main__'`)
+
 ## Version 1.1.1: Grading Support
 
 * Added experimental `modlib` module (unexposed)